feat(keymaps): add end of line shortcut and update whick-key config
In `core/keymaps.lua`, a new shortcut for navigating to the end of the
line is added. The key '9' is now mapped to the end of line command.

In `plugins/helpers/whick-key.lua`, the configuration for mode triggers
is simplified. Now, only the leader key in normal mode triggers the
whick-key helper.