feat(keymaps, copilot): Update keymaps and CopilotChat setup
In `keymaps.lua`, the keymap for "End of line" has been updated to work in
both normal and visual mode. In `copilot.lua`, the CopilotChat setup has been
enhanced with the addition of cmp integration and new key mapping for
"CopilotChat - Prompt actions". Also, the `complete` mapping in CopilotChat
setup has been updated.